Moves event listeners callbacks subscription to views.
