+++ src/view/film-details-view.js
@@ -173,4 +173,14 @@ export default class FilmDetailsView extends AbstractView {
     return getFilmDetailsViewTemplate(this.#film, this.#comments);
   }
 
+  setCloseButtonClickHandler = (callback) => {
+    this._callback.click = callback;
+    this.element.querySelector('.film-details__close-btn').addEventListener('click', this.#closeButtonClickHandler);
+  };
+
+  #closeButtonClickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.click();
+  };
+
 }
